% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bee.R
\name{bee}
\alias{bee}
\title{bee}
\usage{
bee(y, grouping = NULL, data, logtransform = FALSE, xlab = grouping,
  ylab = y, main = if (!is.null(grouping)) {     rbind(ylab, "by", xlab) }
  else {     ylab }, grouplabels = FALSE, color = 1, dotshape = 16,
  boxplot = FALSE, boxplotRange = 1.5, method = "swarm",
  corral = "wrap")
}
\arguments{
\item{y}{Name of the y variable}

\item{grouping}{Name of the grouping variable}

\item{data}{Input dataframe where y and grouping are found}

\item{logtransform}{Should the data be log transformed}

\item{xlab}{X-axis Label}

\item{ylab}{Y-axis Label}

\item{main}{Main Title}

\item{grouplabels}{Label for the grouping variable}

\item{color}{Vector for the dot colors}

\item{dotshape}{Vecor for the dot shapes}

\item{boxplot}{Should a boxplot be plotted along with beeswarm plot?}

\item{boxplotRange}{How long should the whiskers be? Default is 1.5}

\item{method}{Beeswarm plotting method. Default is 'swarm'. Options include: 'swarm', 'center', 'hex', 'square'}

\item{corral}{Further plotting method. Default is 'wrap', Options include: 'none', 'gutter', 'wrap', 'random', 'omit'}
}
\value{
A beeswarm plot
}
\description{
bee is a convenience wrapper for the beeswarm::beeswarm function
}
\examples{
#dat<-data.frame(group=rep(c(1,2,3),40),x1=rnorm(120,18,5))
#names(dat)<-c("group","x1")
#bee(y="x1",
#    grouping="group",
#    data=dat,
#    boxplot=TRUE,
#    xlab="Severity",
#    ylab="Score",
#    main="Some Made Up Data",
#    grouplabels=c("Low","Medium","High"),
#    logtransform=FALSE)
}

