% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandAltman.R
\name{blandAltman}
\alias{blandAltman}
\title{blandAltman}
\usage{
blandAltman(x, y, data, main = NULL, xlab = NULL, ylab = NULL,
  CIBand = FALSE, CIInterval = NULL, lty = "dashed", CIBand2 = FALSE,
  CIInterval2 = NULL, lty2 = "dashed", baseline = "zero",
  standardized = FALSE)
}
\arguments{
\item{x}{Continuous variable}

\item{y}{Continuous variable}

\item{data}{data.frame containing x and y}

\item{main}{Main title}

\item{xlab}{X-axis Label}

\item{ylab}{Y-axis Label}

\item{CIBand}{TRUE/FALSE: Plot a confidence interval}

\item{CIInterval}{What should the interval be at? Default is 95%CI. User input will be at +/- user input.}

\item{lty}{Line type}

\item{CIBand2}{TRUE/FALSE: Plot a second confidence interval}

\item{CIInterval2}{What should the interval be at? Default is 99%CI. User input will be at +/- user input}

\item{lty2}{Line Type for CIBand2}

\item{baseline}{Should there be a 'baseline' line drawn. 'none' = no line, 'zero' = line at 0, 'bias' = line at the level of bias}

\item{standardized}{TRUE/FALSE: Should x and y be standardized?}
}
\value{
A Bland-Altman plot
}
\description{
blandAltman prints a Bland Altman plot
}
\examples{
#NULL
}

