% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFootnote.R
\name{ggFootnote}
\alias{ggFootnote}
\title{ggFootnote}
\usage{
ggFootnote(plotObject, footnote, fileName = NULL, xLocation = "left",
  fontface = "plain", fonttype = "HersheySerif", fontcolor = "black",
  fontsize = 10, hjust = -0.1, vjust = 0.1, width_inch = 5,
  height_inch = 5, dpi = 600)
}
\arguments{
\item{plotObject}{A ggplot2 object}

\item{fileName}{filename to save the output object to. If NULL, then the object will be printed to the plots viewer and not saved.}

\item{xLocation}{x axis location where the text should be added. Default is 'left'. Options are 'left', 'center', or 'right', or a numeric x-axis entry.}

\item{fontface}{Font-face to be used. Default is 'plain', other options include 'bold', 'italic', 'bold.italic', etc}

\item{fonttype}{Font-type to be used. Default is 'HersheySerif'}

\item{fontcolor}{Font color to be used. Default is 'black'}

\item{fontsize}{Fond size to use. Default is 10.}

\item{hjust}{Horizontal fine-adjustment to alignment. Default is -.1.}

\item{vjust}{Vertical fine-adjustment to alignment. Default is .1. Note that if you want the footnote at the top, you can use vjust to allow for gross-adjustment to the vertical alignment as well.}

\item{width_inch}{If fileName is non-NULL, what should the width of the output plot file be.}

\item{height_inch}{If fileName is non-NULL, what should the height of the output plot file be.}

\item{dpi}{If fileName is non-NULL, what should the dpi of the output plot file be.}
}
\value{
A ggplot2/gridExtra object
}
\description{
ggFootnote adds a footnote or, more generally, text around the margins of a ggplot2 object
}
\examples{
#NULL
}

