% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSurv.R
\name{ggSurv}
\alias{ggSurv}
\title{ggSurv}
\usage{
ggSurv(s, CI = "def", plot.cens = T, surv.col = "gg.def",
  cens.col = "red", cens.size = 1, lty.est = 1, lty.ci = 2,
  cens.shape = 3, theme_classic = T, xlab = "Time", ylab = "Survival",
  main = "", legendName = NULL, legendLabels = NULL)
}
\arguments{
\item{s}{suvival object from survdiff(Surv())}

\item{plot.cens}{TRUE/FALSE: Plot censored observations?}

\item{surv.col}{Color for survival curves}

\item{cens.col}{Color for censor ticks}

\item{cens.size}{Size of censor ticks}

\item{lty.est}{Line-type for survival curves}

\item{lty.ci}{Line-type for survival curve confidence intervals}

\item{cens.shape}{Censor tick shapes}

\item{theme_classic}{TRUE/FALSE: Should the theme be set to theme_class()}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{main}{main title}

\item{legendName}{Title for the legend}

\item{legendLabels}{Labels for the legend}
}
\value{
A ggplot2 Kaplan-Meier survival curve object
}
\description{
ggSurv
}
\examples{
#NULL
}

