% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinalccf.R
\name{longitudinalccf}
\alias{longitudinalccf}
\title{longitudinalccf}
\usage{
longitudinalccf(n, y, x, id, data, outputdata = FALSE, singlecor = FALSE,
  plot = TRUE, onlylead = FALSE, main = NULL, sub = NULL, xlab = NULL,
  ylab = NULL, yscale = FALSE)
}
\arguments{
\item{n}{Number of lead/lags to compute up to}

\item{y}{Y variable}

\item{x}{X variable}

\item{id}{id variable, unique across pts.}

\item{data}{data.frame containing all data}

\item{outputdata}{If TRUE, a data.frame containing all lead and lag variables appended to the right}

\item{singlecor}{If TRUE, will output the lead/lag correlations for the specified n separately}

\item{plot}{If TRUE, a plot of all correlations will be drawn}

\item{onlylead}{If TRUE, only leads will be estimated, no lags}

\item{main}{Main title for plot, if plot==TRUE}

\item{sub}{Sub title for plot, if plot==TRUE}

\item{xlab}{X-axis label for plot, if plot==TRUE}

\item{ylab}{Y-axis label for plot, if plot==TRUE}

\item{yscale}{Y scale for plot, if plot==TRUE}
}
\value{
data.frame with new data and/or a plot of correllation coefficients.
}
\description{
A conceptual extension to the ccf() function. Computes corrleation coeffeicients for various leads and lags between an x and y variable, for longitudinal data were observations are nested within an 'id' variable.
}
\examples{
#get<-function(){
#id<-sort(rep(1:200, 50))
#time<-rep(1:50,200)
#main<-(rnorm(10000, 7.4,2.2))
#effect<-time*.85
#y<-main + effect
#return(data.frame(id,time,y))
#}
#df <- get()
#longitudinalccf(n = 15,
#                y = y,
#                 x = time,
#                 id = id,
#                 data = df,
#                 plot=TRUE,
#                 outputdata=TRUE,
#                 singlecor=FALSE,
#                 onlylead=FALSE)
}

