% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPlot.R
\name{multiPlot}
\alias{multiPlot}
\title{multiPlot}
\usage{
multiPlot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Plots to be combined together}

\item{plotlist}{List of plots to be combined together}

\item{file}{Name for the output file}

\item{cols}{Number of columns}

\item{layout}{An x, y coordinate system for laying out the plots}
}
\value{
A ggplot object
}
\description{
multiPlot creates a ggplot object that consists of multiple ggplot subplots
}
\examples{
#NULL
}

