% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multVarPlot.R
\name{multVarPlot}
\alias{multVarPlot}
\title{multVarPlot}
\usage{
multVarPlot(data, varList, grouping, legendTitle = NULL,
  legendLabels = NULL, xlab = NULL, xlabels = NULL, ylab = NULL,
  ylabels = NULL, colourManual = NULL, shapeManual = NULL, title = NULL,
  errorBars = FALSE, errorBarWidth = 0.1, errorBarType = "se")
}
\arguments{
\item{data}{data.frame containing variables for plotting}

\item{varList}{Vector of variable names to plot across the x-axis}

\item{grouping}{Variable name to use for grouping data}

\item{legendTitle}{Title for the legend, given a grouping variable}

\item{legendLabels}{Labels for the legend, given a grouping variable}

\item{xlab}{X-axis label}

\item{xlabels}{X-axis value labels to match to the varList variable names}

\item{ylab}{Y-axis label}

\item{ylabels}{Y-axis value labels}

\item{colourManual}{Manual colour pallet for grouping variable levels}

\item{shapeManual}{Manual shape pallet for grouping variable levels}

\item{title}{Main title for plot}

\item{errorBars}{If TRUE, error bars will be added to the plot}

\item{errorBarWidth}{If errorBars==TRUE, width of the error bars}

\item{errorBarType}{If errorBars==TRUE, error bar type: 'sd', 'se', or 'ci'}
}
\value{
A ggplot2 object
}
\description{
multVarPlot provides a simple interface for plotting multiple columns in a data.frame across common X and Y axes.
}
\examples{
#data(iris)
#l <- multVarPlot(data=iris,
#              varList= c("Sepal.Length", "Sepal.Width",
#                         "Petal.Length", "Petal.Width"),
#              grouping = "Species",
#              colourManual = c("blue", "red", "black"),
#              shapeManual = c(15, 16, 17),
#              errorBars = T,
#              errorBarType = "sd")
#l + theme(legend.position="bottom") +
#   scale_x_discrete("Measurement Type",
#                    labels=c("Sepal.Length"="SL",
#                             "Sepal.Width"="SW",
#                             "Petal.Length"="PL",
#                             "Petal.Width"="PW"))
}

