% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoAxisPlot.R
\name{twoAxisPlot}
\alias{twoAxisPlot}
\title{twoAxisPlot}
\usage{
twoAxisPlot(y1, y2, x1, x2, main, ylab1, ylab2, xlab, ylim1 = NULL,
  ylim2 = NULL, xBreaks = NULL, xLabels = NULL, legendNames)
}
\arguments{
\item{y1}{Vector of y1}

\item{y2}{Vector of y2}

\item{x1}{Vector of x1}

\item{x2}{Vector of x2 (if same x should be used for both, can just specify x1's vector a second time)}

\item{main}{Main Title}

\item{ylab1}{Y-axis label for y1}

\item{ylab2}{Y-axis label for y2}

\item{xlab}{X-axis label}

\item{ylim1}{Upper and lower limits for y1}

\item{ylim2}{Upper and lower limits for y2}

\item{xBreaks}{Breaks for X-axis. If NULL, then all unique levels across x1 and x2 are used}

\item{xLabels}{Labels for X-axis values. If NULL, then actual values are used.}

\item{legendNames}{Names of y1 and y2 for legend}
}
\value{
Plot object printed to plots panel
}
\description{
twoAxisPlot plots two different y vectors to a single plot with two separate y-axes on the left and right of the plot.
}
\examples{
x1 <- c(0, 35, 100, 300)
x2 <- c(0, 100, 300)
y1 <- c(10, 12, 16, 13)
y2 <- c(900, 850, 825)
twoAxisPlot(y1 = y1, y2 = y2, x1 = x1, x2 = x2,
            main = 'Example plot',
            ylab1 = 'Y Variable 1',
            ylab2 = 'Y Variable 2',
            xlab = 'X Variable',
            ylim1 = c(10, 20), ylim2 = c(800, 1000),
            xBreaks = NULL, xLabels = NULL,
            legendNames = c("Variable 1", "Variable 2"))
}

