% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twowaybox.R
\name{twowaybox}
\alias{twowaybox}
\title{twowaybox}
\usage{
twowaybox(data, group1, group2 = NULL, y, ylim = NULL, main = NULL,
  labflip = TRUE)
}
\arguments{
\item{data}{data.frame containing all data}

\item{group1}{name of grouping variable 1}

\item{group2}{name of grouping variable 2}

\item{y}{name of y}

\item{ylim}{y limits}

\item{main}{Main title of plot}

\item{labflip}{If TRUE, group 2 labels will be flipped 90 degrees}
}
\value{
base graph figure
}
\description{
twowaybox provides black and white publication quality box plots for a 1 or 2-way design
}
\examples{
#test.data<-function(){                                                       #
#   factor2<-rbinom(100,1,.5)                                                  #
#   factor21<-rbinom(100,1,.5)                                                 #
#   factor3<-sample(c('Male','Female','Other'),100,replace=TRUE)               #
#   factor31<-sample(c('Catholic','Christian','Atheist'),100,replace=TRUE)     #
#   factor4<-rep(1:4,25)                                                       #
#   factor41<-sample(c('U.S.','Australia','UK','Canada'),100,replace=TRUE)     #
#   factor5<-rep(1:5,20)                                                       #
#   factor51<-sample(c('Caucasian','Af.Amer','Asian Am.','Latin Am.',          #
#                      'Canadian'),100,replace=TRUE)                           #
#   y<-rnorm(100,53,19)                                                        #
#   tester<-data.frame(factor2,factor21,factor3,factor31,factor4,factor41,     #
#                      factor5,factor51,y)                                     #
#   return(tester)                                                             #
#}                                                                            #
#okaydata <- test.data()                                                        #
#twowaybox(labflip=FALSE,data=okaydata,group1=factor5,group2=factor2,y=y)
#twowaybox(labflip=FALSE,data=okaydata,group1=factor5,y=y)
#twowaybox(labflip=FALSE,data=okaydata,group1=factor3,y=y)
#twowaybox(labflip=FALSE,data=okaydata,group1=factor2,y=y)
}

